# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Backend API
  description: Spring Boot Backend Application REST API
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local Development Server
  - url: https://api.example.com
    description: Production Server
security:
  - bearerAuth: []
paths:
  /api/auth/signin:
    post:
      tags:
        - Authentication
      summary: Sign in user
      description: Authenticate user and return JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Sign up user
      description: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/products:
    get:
      tags:
        - Product Management
      summary: Get all products
      description: Retrieve paginated list of all products
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Sort by field
          schema:
            type: string
            default: id
        - name: sortDir
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: search
          in: query
          description: Search term
          schema:
            type: string
        - name: category
          in: query
          description: Category filter
          schema:
            type: string
        - name: minPrice
          in: query
          description: Minimum price
          schema:
            type: number
            format: decimal
        - name: maxPrice
          in: query
          description: Maximum price
          schema:
            type: number
            format: decimal
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedProducts'

    post:
      tags:
        - Product Management
      summary: Create product
      description: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/products/{id}:
    get:
      tags:
        - Product Management
      summary: Get product by ID
      description: Retrieve a specific product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Product Management
      summary: Update product
      description: Update an existing product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Product Management
      summary: Delete product
      description: Soft delete a product (mark as inactive)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product deleted successfully
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/products/low-stock:
    get:
      tags:
        - Product Management
      summary: Get low stock products
      description: Retrieve products with low stock
      parameters:
        - name: threshold
          in: query
          description: Stock threshold
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Low stock products retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/products/{id}/stock:
    patch:
      tags:
        - Product Management
      summary: Update product stock
      description: Update product stock quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 0
              required:
                - quantity
      responses:
        '200':
          description: Stock updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: admin123
    SignUpRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: newuser
        email:
          type: string
          format: email
          maxLength: 100
          example: newuser@example.com
        password:
          type: string
          format: password
          minLength: 6
          maxLength: 40
          example: password123
        role:
          type: array
          items:
            type: string
            enum: [admin, mod, user]
          example: ["user"]
    JwtResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          type: string
          example: Bearer
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@example.com
        roles:
          type: array
          items:
            type: string
          example: ["ROLE_ADMIN"]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully!
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Laptop
        description:
          type: string
          example: High-performance laptop for professionals
        price:
          type: number
          format: decimal
          example: 1299.99
        quantity:
          type: integer
          example: 50
        category:
          type: string
          example: Electronics
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: 2023-12-19T10:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-12-19T10:00:00Z
    ProductRequest:
      type: object
      required:
        - name
        - price
        - quantity
      properties:
        name:
          type: string
          maxLength: 100
          example: New Product
        description:
          type: string
          maxLength: 500
          example: Product description
        price:
          type: number
          format: decimal
          minimum: 0.01
          example: 99.99
        quantity:
          type: integer
          minimum: 0
          example: 100
        category:
          type: string
          maxLength: 50
          example: Electronics
...
    PagedProducts:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
          example: 100
        totalPages:
          type: integer
          example: 10
        last:
          type: boolean
          example: false
        first:
          type: boolean
          example: true
        numberOfElements:
          type: integer
          example: 10
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
    Pageable:
      type: object
      properties:
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 10
        offset:
          type: integer
          format: int64
          example: 0
        paged:
          type: boolean
          example: true
        unpaged:
          type: boolean
          example: false
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
          example: true
        unsorted:
          type: boolean
          example: false
        empty:
          type: boolean
          example: false
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: 2023-12-19T10:00:00Z
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Validation failed
        path:
          type: string
          example: /api/users
        errors:
          type: object
          additionalProperties:
            type: string
          example:
            username: Username is required
            email: Email format is invalid