# backend/src/main/java/com/example/backend/dto/MessageResponse.java
package com.example.backend.dto;

public class MessageResponse {
    
    private String message;
    
    public MessageResponse(String message) {
        this.message = message;
    }
    
    public String getMessage() {
        return message;
    }
    
    public void setMessage(String message) {
        this.message = message;
    }
}